requirements.txt
opencv-python>=4.1.1.26
PyYAML>=5.3.1
PyAutoGUI>=0.9.50
numpy>=1.17.2
wxPython>=4.1.0